ref (tree): separated funcs & index file moved presentFilesAsTree & printTree to utils
